Add pivot.service as a host API
Add a new systemd service that simply executes `pivot`. But note that we
don't have a `pivot.path` unit to auto-trigger it: this needs to be
triggered explicitly by whatever drives pivot. Making it a service and
standizing on a path in `/etc` means that we can trigger a pivot on boot
from Ignition (patch to preset enable the service is pending), as well
as from the MCD by directly starting the unit.

By running it directly from the host context rather than e.g. directly
by the MCD, we also avoid potential issues with SELinux as described in:
https://github.com/openshift/machine-config-operator/issues/314